- HeuristikTest package in Main package gemerged und dann gelöscht - möglichst gut für merge mit develop aufgeräumt

+++ src/main/java/Main/Client.java
@@ -9,13 +9,12 @@ public class Client {
     private boolean isRunning;
     private boolean bomben;
     private static long max = Long.MIN_VALUE, min = Long.MAX_VALUE;
-   // short[] zug= new short[3];
+    // int[] zug= new int[3];
 
 
     public Client() throws IOException {
 
     }
-
 
     public void netzwerk(int p) throws IOException {
         String ip = "127.0.0.1"; // localhost
@@ -40,7 +39,7 @@ public class Client {
         DynamischeHeuristik heuristik = new DynamischeHeuristik(Spiel, 1);
         System.out.println(heuristik);
 
-        while(isRunning) {
+        while (isRunning) {
             empfangeNachricht(socket);
         }
 
@@ -62,20 +61,20 @@ public class Client {
 
         //ArrayList <Byte> nachricht = new ArrayList<>();
 
-        socketInputStream.read(stream,0,5);
+        socketInputStream.read(stream, 0, 5);
 
 
         byte art = stream[0];
         int laenge = getlaenge(stream);
-        char nachricht[] ;
+        char nachricht[];
         nachricht = new char[laenge];
         byte message[] = new byte[laenge];
-        socketInputStream.read(message,0,laenge);
+        socketInputStream.read(message, 0, laenge);
 
-        for(int i = 0; i <laenge;i++){
-            nachricht[i] = (char)message[i];
+        for (int i = 0; i < laenge; i++) {
+            nachricht[i] = (char) message[i];
         }
-       // System.out.println(art);
+        // System.out.println(art);
         switch (art) {
             case 2:
                 return String.valueOf(nachricht);
@@ -86,13 +85,13 @@ public class Client {
                 break;
             case 4:
                 char[][][] Spielfeld = Spiel.getSpielfeld();
-                if(bomben){
+                if (bomben) {
                     Spiel.gueltigeBombZuege();
                     //Todo Sinnvolle Zugauswahl
 
-                    for (short y = 0; y < Spiel.getHoehe(); y++) {
-                        for (short x = 0; x < Spiel.getBreite(); x++) {
-                            short[] zug = new short[3];
+                    for (int y = 0; y < Spiel.getHoehe(); y++) {
+                        for (int x = 0; x < Spiel.getBreite(); x++) {
+                            int[] zug = new int[3];
                             if (Spielfeld[x][y][1] == 'B') {
                                 zug[0] = x;
                                 zug[1] = y;
@@ -104,21 +103,21 @@ public class Client {
 
                 } else {
                     //Todo Sinnvolle Zugauswahl
-                    short[] zug = new short[3];
+                    int[] zug = new int[3];
                     long start, ende, gesamt;
                     start = System.nanoTime();
                     zug = Spiel.sucheZug(5, Spielernummer);
                     ende = System.nanoTime();
-                    gesamt = ende-start;
-                    if(max < gesamt) {
+                    gesamt = ende - start;
+                    if (max < gesamt) {
                         max = gesamt;
                     }
-                    if(min > gesamt) {
+                    if (min > gesamt) {
                         min = gesamt;
                     }
-                    System.out.println("Zeit: "+gesamt/1000000.0+" ms");
-                    System.out.println("Maximale Zeit: "+max/1000000.0+" ms");
-                    System.out.println("Minimale Zeit: "+min/1000000.0+" ms");
+                    System.out.println("Zeit: " + gesamt / 1000000.0 + " ms");
+                    System.out.println("Maximale Zeit: " + max / 1000000.0 + " ms");
+                    System.out.println("Minimale Zeit: " + min / 1000000.0 + " ms");
                     sendeZug(zug, socket);
 
                     return "";
@@ -127,25 +126,25 @@ public class Client {
                 break;
 
             case 6:
-                short x = (short)((short)message[0]<<8);
+                int x = message[0] << 8;
                 x += message[1];
 
-                short y = (short)((short)message[2]<<8);
+                int y = message[2] << 8;
                 y += message[3];
                 byte sonderfeld = message[4];
                 byte spieler = message[5];
                 Spielfeld = Spiel.getSpielfeld();
 
-                if(bomben){
-                     Spiel.bombZug(x,y);
-                } else if (spieler == Spielernummer){
+                if (bomben) {
+                    Spiel.bombZug(x, y);
+                } else if (spieler == Spielernummer) {
                     Spiel.ganzerZug(spieler, x, y);
                 } else {
                     int anzahlsteine = Spiel.getUeberschreibsteine();
                     Spiel.ganzerZug(spieler, x, y);
                     Spiel.setUeberschreibsteine(anzahlsteine);
                 }
-               break;
+                break;
             case 7:
                 System.exit(-1);
             case 8:
@@ -160,7 +159,7 @@ public class Client {
 
     }
 
-    public void sendeZug(short[] zug,java.net.Socket socket)throws IOException{
+    public void sendeZug(int[] zug, java.net.Socket socket) throws IOException {
         byte laenge = 5;
         byte nachricht[];
         byte zuSendendeNachricht[];
@@ -168,30 +167,29 @@ public class Client {
         zuSendendeNachricht = new byte[laenge + 5];
         zuSendendeNachricht[0] = 5;
         zuSendendeNachricht[4] = laenge;
-        short x = zug[0];
-        short y = zug[1];
-        byte sonderfeld = (byte)zug[2];
+        int x = zug[0];
+        int y = zug[1];
+        byte sonderfeld = (byte) zug[2];
 
-        zuSendendeNachricht[6] = (byte)(x);
-        zuSendendeNachricht[5] = (byte)((x >> 8));
+        zuSendendeNachricht[6] = (byte) (x);
+        zuSendendeNachricht[5] = (byte) ((x >> 8));
 
-        zuSendendeNachricht[8] = (byte)(y);
-        zuSendendeNachricht[7] = (byte)(y >> 8);
+        zuSendendeNachricht[8] = (byte) (y);
+        zuSendendeNachricht[7] = (byte) (y >> 8);
         zuSendendeNachricht[9] = sonderfeld;
 
-        schreibeNachricht(socket,zuSendendeNachricht);
+        schreibeNachricht(socket, zuSendendeNachricht);
 
     }
 
-
     private int getlaenge(byte[] bytes) {
 
-            int value = 0;
-            for (int i = 0; i < 4; i++) {
-                int shift = (4 - 1 - i) * 8;
-                value += (bytes[i+1] & 0x000000FF) << shift;
-            }
-            return value;
+        int value = 0;
+        for (int i = 0; i < 4; i++) {
+            int shift = (4 - 1 - i) * 8;
+            value += (bytes[i + 1] & 0x000000FF) << shift;
+        }
+        return value;
 
     }
 }
